formatHouseNumbers
This needs to be stepped through with the related workbook.  It's not the most efficient version because I wrote it quickly to get the job done but it contains some useful functions like trimming unwanted characters from a string, a dictionary of terms to remove from a range and formatting for specific events.