Control position of Userform
When put into the module calling the Userform, this will determine where the center of the Excel window is and open the userform at that position.  Note that this uses the center of the Excel window so only if the window is maximized, will it find the center of the monitor.