Get a date from a Work Week number
This script calculates the starting date of a work week. We had two companies that shared data but one worked from a calendar year and the other a fiscal year that started in September so we needed an easy way to translate them. This is the "push button" script developed.