VBA To Run a Third Party Program
This was a very, very well used script.  A customer had created their own in-house software to collect data but it lacked the charting and graphics of Excel.  So we developed this script to open their software, run a pre-configured report, enter our own variables then save the output as a Tab Separated Value file.