Create vbaPause
This script is used to pause the macro while another program creates a file.  If there are a series of files being created, make sure this script looks for the last one created.  Before calling this script, you should make sure the file(s) does not already exist in the directory.  In the example above, the script looks for the existence of the files every 10 seconds.